Store multiple scans in one database.
This is some old code that I found that I never got aroun to
publishing as a tarball. I have ported the code to
PyObject+Introspection, though it needs this fix in PyObject:
https://bugzilla.gnome.org/show_bug.cgi?id=647272

There is also some nasty hardcoding of the sources.list.
That should be in a single-row table instead.